+++ src/Snap.Hutao/Snap.Hutao/Service/Game/Package/Advanced/GamePackageOperationContext.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license.
 
 using Snap.Hutao.Core.DependencyInjection.Abstraction;
-using Snap.Hutao.Web.Hoyolab.HoyoPlay.Connect.Branch;
 using Snap.Hutao.Web.Hoyolab.HoyoPlay.Connect.ChannelSDK;
 using System.IO;
 
@@ -16,7 +15,7 @@ internal readonly struct GamePackageOperationContext
     public readonly SophonDecodedBuild LocalBuild;
     public readonly SophonDecodedBuild RemoteBuild;
     public readonly GameChannelSDK? GameChannelSDK;
-    public readonly string ExtractOrGameDirectory;
+    public readonly string EffectiveGameDirectory;
     public readonly string EffectiveChunksDirectory;
 
     public GamePackageOperationContext(
@@ -34,7 +33,7 @@ internal readonly struct GamePackageOperationContext
         LocalBuild = localBuild;
         RemoteBuild = remoteBuild;
         GameChannelSDK = gameChannelSDK;
-        ExtractOrGameDirectory = extractDirectory ?? gameFileSystem.GetGameDirectory();
+        EffectiveGameDirectory = extractDirectory ?? gameFileSystem.GetGameDirectory();
 
         EffectiveChunksDirectory = kind is GamePackageOperationKind.Verify
             ? Path.Combine(gameFileSystem.GetChunksDirectory(), "repair")
