+++ src/Snap.Hutao/Snap.Hutao/Service/Game/Package/Advanced/GamePackageOperationKind.cs
@@ -10,6 +10,6 @@ internal enum GamePackageOperationKind
     Update,
     Predownload,
 
-    ExtractBlk,
-    ExtractExe,
+    ExtractBlocks,
+    ExtractExecutable,
 }
