+++ src/Snap.Hutao/Snap.Hutao/Service/Game/Package/Advanced/PackageOperation/GamePackageInstallOperation.cs
@@ -1,0 +1,31 @@
+// Copyright (c) DGP Studio. All rights reserved.
+// Licensed under the MIT license.
+
+using System.IO;
+
+namespace Snap.Hutao.Service.Game.Package.Advanced.PackageOperation;
+
+[Injection(InjectAs.Transient, typeof(IGamePackageOperation), Key = GamePackageOperationKind.Install)]
+internal sealed class GamePackageInstallOperation : GamePackageOperation
+{
+    public override async ValueTask ExecuteAsync(GamePackageServiceContext context)
+    {
+        SophonDecodedBuild remoteBuild = context.Operation.RemoteBuild;
+        int totalChunksCount = remoteBuild.TotalChunks;
+        long totalBytes = remoteBuild.TotalBytes;
+
+        InitializeDuplicatedChunkNames(context, remoteBuild.Manifests.SelectMany(m => m.ManifestProto.Assets.SelectMany(a => a.AssetChunks)));
+
+        context.Progress.Report(new GamePackageOperationReport.Reset(SH.ServiceGamePackageAdvancedInstalling, totalChunksCount, totalBytes));
+
+        await context.Operation.Asset.InstallAssetsAsync(context, remoteBuild).ConfigureAwait(false);
+        await context.Operation.Asset.EnsureChannelSdkAsync(context).ConfigureAwait(false);
+
+        await PrivateVerifyAndRepairAsync(context, remoteBuild, totalBytes, totalChunksCount).ConfigureAwait(false);
+
+        if (Directory.Exists(context.Operation.EffectiveChunksDirectory))
+        {
+            Directory.Delete(context.Operation.EffectiveChunksDirectory, true);
+        }
+    }
+}
